DevTools: Fix "uncaught promise rejection" when loading frontend in hosted mode
The core issue is in the way the following chain is processed:

P1 -> thenOrCatch -> P2 -> thenOrCatch

Given that P1 and P2 wrapping async XHR's, they resolve/reject
as the xhrs finish (and XHR's are launched as the promises were created).

If P2 is rejected before P1 finished its loading, it cannot proceed to
its catch handler - it ends the microtask in rejected state and waits for
the P1 to fullfill. This results in the "unhandled promise rejection"

This patch also minimizes amount of tasks created to process scripts 
loading promise chain.

Review URL: https://codereview.chromium.org/691833002

git-svn-id: svn://svn.chromium.org/blink/trunk@184689 bbb929c8-8fbe-4397-9dbb-9b2b20218538