Devirtualize animatedLocalTransform() and rename it for clarity
This patch devirtualizes animatedLocalTransform and renames it
calculateAnimatedLocalTransform to emphasize that it is not cheap.

SVGGraphicsElement::animatedLocalTransform was virtual to support
SVGTextElement's special transform code. The text-specific
transform code is only in an infrequently-taken codepath
(style && style->hasTransform()) so we can actually improve performance
by devirtalizing and combining the two transform codepaths. The real
benefit is code simplification though.

Secondly, I've renamed this function to make it clear that it is not
a simple lookup.

BUG=429551

Review URL: https://codereview.chromium.org/693423002

git-svn-id: svn://svn.chromium.org/blink/trunk@184845 bbb929c8-8fbe-4397-9dbb-9b2b20218538