Remove the HasOwnBackingButPaintsIntoAncestor compositing state.
The referenced bug is caused by a composited descendant that has an opacity
animation not having the correct background painted for its ancestor that it needs
to blend with.

However, HasOwnBackingButPaintsIntoAncestor is also just an optimization to reduce
the number of composited layers that draw content, in cases where we think we can
get away with it. (e.g. only existing to pass along preserves-3d state).

In exchange for these optimizations, the feature introduces significant complexity,
and has caused a number of bugs. Let's just kill it.

BUG=418521

Review URL: https://codereview.chromium.org/708213002

git-svn-id: svn://svn.chromium.org/blink/trunk@185007 bbb929c8-8fbe-4397-9dbb-9b2b20218538