Use different icons for the overlay and non-overlay cast buttons
We want the overlay cast icon to be visible on all backgrounds, so its
background has to be at least semi-opaque. The non-overlay cast icon,
however, should have the same background as the media controls, which
is easiest to achive by making its background transparent. As such we
need different icons for the two cases.

Note that this also needs a Chromium CL for the new versions of the
icons.

BUG=426373

Review URL: https://codereview.chromium.org/716613002

git-svn-id: svn://svn.chromium.org/blink/trunk@185125 bbb929c8-8fbe-4397-9dbb-9b2b20218538