Rename HTMLInputElement::imageLoader() to ensureImageLoader()
Make clear that imageLoader is potentially constructing the
ImageLoader by renaming to ensureImageLoader. Also rename
hasImageLoader to imageLoader and make it return a pointer,
so it can be used both to check if the ImageLoader is created
as well as use it if not null.

http://trac.webkit.org/changeset/175764

Review URL: https://codereview.chromium.org/715633005

git-svn-id: svn://svn.chromium.org/blink/trunk@185143 bbb929c8-8fbe-4397-9dbb-9b2b20218538