ResourceFetcher::requestPreload should set its requestContext
This CL ensures ResourceFetcher::requestPreload to request with
an appropriate context, RequestContextPrefetch.
This CL fixes the issue that ASSERT in
MixedContentChecker::contextTypeFromContext is hit when preloader
is invoked.

This is to prepare for CL http://crrev.com/673603002 , which makes
preloader to be invoked more often.

TESTS=http/tests/security/mixedContent
BUG=421300

Review URL: https://codereview.chromium.org/711213002

git-svn-id: svn://svn.chromium.org/blink/trunk@185169 bbb929c8-8fbe-4397-9dbb-9b2b20218538