Early return from WebPluginContainerImpl::scriptableObject if the plugin gets destroyed
WebPluginContainerImpl::scriptableObject calls into WebPlugin::scriptableObject
which (in the PepperWebPluginImpl implementation) sends synchronous messages
to the plugin which can result in re-entrancy. This re-entrancy may cause the
plugin to be destroyed, which sets m_webPlugin to null in
WebPluginContainerImpl::dispose. We need to check for a null value to avoid
dereferencing a null pointer.

BUG=423263

Review URL: https://codereview.chromium.org/715073002

git-svn-id: svn://svn.chromium.org/blink/trunk@185182 bbb929c8-8fbe-4397-9dbb-9b2b20218538