Expose whether a touch event may cause scrolling if uncanceled
Currently, there are several parts of the touch pipeline that need to
know whether a touch event may cause scrolling. In particular, touchmove
events within a platform-specific slop region will not induce
scrolling, and such events are suppressed by the TouchEventQueue when
the touchstart is not prevented. The TouchEventQueue uses a slop region
constant to perform this suppression.

However, there are no guarantees that this constant is the same as that
used in gesture detection. With devtools touch emulation, it should be
possible to vary the slop region when emulating different devices, but
the duplicated slop region code in the TouchEventQueue makes this
difficult.

This solution tags each touch event with a bit indicating whether
the event may cause scrolling, allowing any listener or consumer of the
touch stream to better reason about the touch event's default action.
This tagging will be wired up in the corresponding Chromium change:
https://codereview.chromium.org/718153002

Also remove the IsLastInputEventForCurrentVSync flag as it's no
longer used.

BUG=425586

Review URL: https://codereview.chromium.org/702883004

git-svn-id: svn://svn.chromium.org/blink/trunk@185244 bbb929c8-8fbe-4397-9dbb-9b2b20218538