IDL: Improve overload resolution for methods with variadic arguments
When calculating "optionality list" for a method (as part of calculating
the effective overload set), treat variadic arguments as optional instead
as required.

WebIDL defines the optionality list as a list of "required", "optional"
and "variadic", whereas we implement it as a list of booleans. A list of
booleans is enough to achieve correct behavior, assuming both "optional"
and "variadic" are considered to be optional (true).

BUG=293561

Review URL: https://codereview.chromium.org/723013003

git-svn-id: svn://svn.chromium.org/blink/trunk@185289 bbb929c8-8fbe-4397-9dbb-9b2b20218538