Timeline: do not imply event.thread.target is the main target
We used to rely on event.thread.target in multiple places where
target was needed. These places mostly imply that this was a main
target (i.e. using it to get DOMModel & co), while it can now be a
different (or null) target. Let's keep main target associated with
timeline in Timeline model (and perhaps return null for recorded
timelines later) and plumb it where it is necessary.

BUG=

Review URL: https://codereview.chromium.org/717243002

git-svn-id: svn://svn.chromium.org/blink/trunk@185294 bbb929c8-8fbe-4397-9dbb-9b2b20218538