Fix 8.1 sdk detection defines
Checking the existence of _WIN32_WINNT_WINBLUE is incorrect, rather
we should check that the value of _WIN32_WINNT is the desired value
as that's the level that the headers are exposing.

Also, move #include of windows.h up above check for include of
versionhelpers.h, otherwise the defines are changed mid-file which
causes problems.

Also, fix typo in spelling of "condition".

(Follow up to https://codereview.chromium.org/39693002)

R=yosin@chromium.org, tkent@chromium.org
BUG=311488, 432748

Review URL: https://codereview.chromium.org/732673002

git-svn-id: svn://svn.chromium.org/blink/trunk@185390 bbb929c8-8fbe-4397-9dbb-9b2b20218538