Yield parser before scripts if a lot of DOM changes were made
Yield the main parser thread when processing chunks from the background
parser if a script block is about to be started and a large number of
DOM elements have been parsed in the current session.

This significantly helps initial render (6-8% in testing) at a slight
cost to DOM Content Loaded (~2% in testing) generally but has a large
benefit in particular for sites that put scripts at the end of their
document (which is a common best-practice).

BUG=427757

Review URL: https://codereview.chromium.org/741493004

git-svn-id: svn://svn.chromium.org/blink/trunk@185567 bbb929c8-8fbe-4397-9dbb-9b2b20218538