Oilpan: Remove an isOrphaned() check from each tracing call
The goal of this CL is to remove the heapPage->isOrphaned() check from popAndInvokeCallback(). The isOrphaned() check was needed before this CL for the following reasons:

a) A conservative GC can find a false-positive pointer to an orphaned heap. We shouldn't trace the false-positive pointer.

b) Due to a programming error, there can be a dangling pointer from a live thread heap to an orphaned heap. We shouldn't trace the dangling pointer.

Regarding a), we can remove the isOrphaned check by moving the check to Heap::lookup(). By having Heap::lookup() return false for a false-positive pointer to an orphaned heap, we can prevent the false-positive pointer from being pushed to the marking stack. Thus we can remove the isOrphaned() check from popAndInvokeCallback().

Regarding b), we can remove the isOrphaned check as follows:

1) When shutting down a thread, move all pages of the thread heap to an orphaned heap. Zap all the pages in the orphaned heap.

2) When a next GC happens, just normally trace all pointers without the isOrphaned() check. If there is a dangling pointer to the orphaned heap (i.e., if there is a programming error), the tracing will just crash. This is an acceptable behavior because we shouldn't have any dangling pointer in the code base.

3) When the GC finishes, remove all the pages from the orphaned heaps.

This CL removes HeapTest.CrossThreadPointerToOrphanedPage because it is testing a behavior where we have a dangling pointer from a live thread heap to a dead thread heap. The test starts crashing after this CL.

BUG=420515

Review URL: https://codereview.chromium.org/731993002

git-svn-id: svn://svn.chromium.org/blink/trunk@185785 bbb929c8-8fbe-4397-9dbb-9b2b20218538